+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# (next major)
+- **Breaking**: Renamed AutoreleasingCache.grab() to load() to conform to FilesystemStorable.
+- **Breaking**: Renamed AutoreleasingCache.clear() to purgeAll() to conform to FilesystemStorable.
+- Added FilesystemStorable, a mixin to standardize classes that load & store from the filesystem.
+- AutoreleasingCache now conforms to FilesystemStorable.
+- **Breaking**: Removed AutoreleasingCache.storeFromDisk(), use load().
+- **Breaking**: Removed AutoreleasingCache.storeFromDiskWithFalsePath(), use load() in combination with rename().
+- **Breaking**: Removed AutoreleasingCache.replace(), use purge() then store().
+
 # 2.0.1
 - Export HTTPStatusCode and ServerRunningException (should've been exposed in 2.0.0)